+++ src/tests/potentials.cc
@@ -465,7 +465,6 @@ static ExperimentParameters default_parameters_vdf(
       true,   // can decay initial particles
       DileptonBremsPionFormFactor::Off,  // no form factor
       SpinInteractionType::Off,          // no spin interactions
-      std::nullopt                       // use monash tune, not known
   };
 }